Add Base Multiple Table File Source
