+++ seatunnel-connectors-v2/connector-file/connector-file-local/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/local/source/LocalFileSource.java
@@ -18,71 +18,18 @@
 package org.apache.seatunnel.connectors.seatunnel.file.local.source;
 
 import org.apache.seatunnel.api.configuration.ReadonlyConfig;
-import org.apache.seatunnel.api.source.Boundedness;
-import org.apache.seatunnel.api.source.SeaTunnelSource;
-import org.apache.seatunnel.api.source.SourceReader;
-import org.apache.seatunnel.api.source.SourceSplitEnumerator;
-import org.apache.seatunnel.api.source.SupportColumnProjection;
-import org.apache.seatunnel.api.source.SupportParallelism;
-import org.apache.seatunnel.api.table.catalog.CatalogTable;
-import org.apache.seatunnel.api.table.type.SeaTunnelRow;
-import org.apache.seatunnel.connectors.seatunnel.file.config.BaseFileSourceConfig;
 import org.apache.seatunnel.connectors.seatunnel.file.config.FileSystemType;
 import org.apache.seatunnel.connectors.seatunnel.file.local.source.config.MultipleTableLocalFileSourceConfig;
-import org.apache.seatunnel.connectors.seatunnel.file.source.reader.MultipleTableFileSourceReader;
-import org.apache.seatunnel.connectors.seatunnel.file.source.split.FileSourceSplit;
-import org.apache.seatunnel.connectors.seatunnel.file.source.split.MultipleTableFileSourceSplitEnumerator;
-import org.apache.seatunnel.connectors.seatunnel.file.source.state.FileSourceState;
+import org.apache.seatunnel.connectors.seatunnel.file.source.BaseMultipleTableFileSource;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class LocalFileSource
-        implements SeaTunnelSource<SeaTunnelRow, FileSourceSplit, FileSourceState>,
-                SupportParallelism,
-                SupportColumnProjection {
-
-    private final MultipleTableLocalFileSourceConfig multipleTableFileSourceConfig;
+public class LocalFileSource extends BaseMultipleTableFileSource {
 
     public LocalFileSource(ReadonlyConfig readonlyConfig) {
-        this.multipleTableFileSourceConfig = new MultipleTableLocalFileSourceConfig(readonlyConfig);
+        super(new MultipleTableLocalFileSourceConfig(readonlyConfig));
     }
 
     @Override
     public String getPluginName() {
         return FileSystemType.LOCAL.getFileSystemPluginName();
-    }
-
-    @Override
-    public Boundedness getBoundedness() {
-        return Boundedness.BOUNDED;
-    }
-
-    @Override
-    public List<CatalogTable> getProducedCatalogTables() {
-        return multipleTableFileSourceConfig.getFileSourceConfigs().stream()
-                .map(BaseFileSourceConfig::getCatalogTable)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public SourceReader<SeaTunnelRow, FileSourceSplit> createReader(
-            SourceReader.Context readerContext) {
-        return new MultipleTableFileSourceReader(readerContext, multipleTableFileSourceConfig);
-    }
-
-    @Override
-    public SourceSplitEnumerator<FileSourceSplit, FileSourceState> createEnumerator(
-            SourceSplitEnumerator.Context<FileSourceSplit> enumeratorContext) {
-        return new MultipleTableFileSourceSplitEnumerator(
-                enumeratorContext, multipleTableFileSourceConfig);
-    }
-
-    @Override
-    public SourceSplitEnumerator<FileSourceSplit, FileSourceState> restoreEnumerator(
-            SourceSplitEnumerator.Context<FileSourceSplit> enumeratorContext,
-            FileSourceState checkpointState) {
-        return new MultipleTableFileSourceSplitEnumerator(
-                enumeratorContext, multipleTableFileSourceConfig, checkpointState);
     }
 }
